small cleanups (#7)
# What :computer: 
* Adjusted the particles configuration 
* cleaned up some things missed when testing nuxt content
* adjusted prettier ignore


# Why :hand:
* cleaner templates and particles configuration is now in a separate
file
* components made for Nuxt Content need to be placed in
`/components/content/` folder. This accidentally got covered by the
prettier ignore file. The purpose of the prettier ignore for markdown
content is so that we don't have prettier formatting non-standard
markdown content (like vue component syntax) and breaking them. This is
why markdown is ignored for now on prettier auto formatting.