feat: create code switcher tab (#8)
# What :computer: 
* Create a component that tabs content for display
* Create a component that displays partial markdown files for re-use
across documents

# Why :hand:
* Breaks down complex instructions that will be hidden in tabs into
partials for easier editing.
* Makes it easier to re-use content versus re-writing content.

# Evidence :camera:
![Screenshot 2024-03-27 at 3 21
48 PM](https://github.com/matter-labs/zksync-docs/assets/812331/2ff1eb91-fa1f-4231-82de-51f61dd51702)

Example set up in the Getting Started page to show how it can be used.
NuxtUI has a UTabs component that can be used plainly if we didn't want
to use this. This new component provides a way to have tabbed content
that can be loaded from partial markdown files to help split up content
and make editing potentially easier.

A future feature to implement in this component would be to have a way
to track which tab is actively selected and update similar tabs to the
same selected tab. Also store this selected tab in localStorage
potentially so that the preferred tab for the user is always selected
across the site.