feat: create parser for zksync info (#12)
# What :computer: 
* Create a parser that replaces tags in markdown content with values
defined in a _zksync.json file.

# Why :hand:
* For very important content, we don't want to copy/past and rewrite
things all over the docs.

# How it works

Content is structured in a json file under `content/_zksync.json`. Tags
are wrapped in `%%` and are prefixed with `zk_`. The path to the value
in the json file is represented with underscores instead of a dot.

Let's say we want to put the Mainnet block explorer URL into a page
somewhere. It's in the JSON file as `mainnet.block_explorer_url`. The
tag for it will be `%%zk_mainnet_block_explorer_url%%`.