chore: setup redirects for old urls (#99)
# Description

We have a number of links out there that exist that point to pages that
will eventually no longer exist when the new Docs site is released.
These links need to be set up with rewrites to ensure visitors still get
to the pages they're trying to get to. All existing tutorials that are
in Code Community have rewrites directed to them. Any tutorials that do
not exist in the Code Community will be redirected to the total list of
tutorials on Code Community.

Eventually these public links will be updated to point to the new URLs.

Part of
[DEVRL-730](https://linear.app/matterlabs/issue/DEVRL-730/[launch]-handle-rewrites-for-old-urls)