chore: add concurrency rules to workflows (#102)
# Description

Add concurrency rules to workflows for staging and production deploy
workflows. This cancels any currently running workflow when a new
workflow is set to run. For example when two PRs are merged to main
around the same time, one will cancel out the other and only one
workflow is run to deploy.
This makes it so that we don't have a queue of build and deploys to run,
cutting down time and resources.

Fix yaml formatting errors with github issue templates. They do not list
in Create Issues if format is incorrect.