docs: explain windows 10 alg limitation
Because windows 10 only supports RS256, it's not supported by our ES256
verifier, so until we add an RS256 verifier this platform won't work.
We've attempted to limit these sign-ups via the browser, but it doesn't
appear to stop users from choosing that option. Documenting it is the
next best thing, but we'll probably add a more explicit error next.